PWGGA/GammaConv: Add radius dependence to meson-jet corr.
- Histograms added to study the meson production as function of jet pt,
meson pt, and the radius between jet and meson
- As this needs an analyisis in three dimensions, a six dimensional
unfolding matrix is needed. This is realized in a new NDimensional
matrix handler
- Defintion of z changed from a simple division of the pts: z =
ptMeson/ptJet to the scalar product of jet and meson momentum normalized
to the squared lenght of the jet momentum. This allows the definition to
be independent of the jet direction. As the Fragmentation should probe
the momentum fraction of the meson to the jet, it should be in the
reference frame of the jet
- px, py and pz of the true jet were added to the ConvJet task in order
to properly calculate z